feat:refactor code

+++ src/AddLiquidityLisk.sol
@@ -9,19 +9,19 @@ interface IUniswapV2Router {
     function addLiquidity(
         address tokenA,
         address tokenB,
-        uint amountADesired,
-        uint amountBDesired,
-        uint amountAMin,
-        uint amountBMin,
+        uint256 amountADesired,
+        uint256 amountBDesired,
+        uint256 amountAMin,
+        uint256 amountBMin,
         address to,
-        uint deadline
-    ) external returns (uint amountA, uint amountB, uint liquidity);
+        uint256 deadline
+    ) external returns (uint256 amountA, uint256 amountB, uint256 liquidity);
 }
 
 contract UniswapTest {
     address userAddr = msg.sender;
-    uint32 chainId = 11155111;  // Sepolia chain ID
-    address target = 0xeE567Fe1712Faf6149d80dA1E6934E354124CfE3;  // Uniswap Router address on Sepolia
+    uint32 chainId = 11155111; // Sepolia chain ID
+    address target = 0xeE567Fe1712Faf6149d80dA1E6934E354124CfE3; // Uniswap Router address on Sepolia
 
     // Token addresses on Sepolia
     address constant USDC = 0x7Fc21ceb0C5003576ab5E101eB240c2b822c95d2;
@@ -29,8 +29,8 @@ contract UniswapTest {
 
     function addLiquidity() external {
         // Calculate minimum amounts (e.g., 95% of desired amounts)
-        uint256 usdcAmount = 2 * 10**6;  // 2 USDC
-        uint256 daiAmount = 2 * 10**18;  // 2 DAI
+        uint256 usdcAmount = 2 * 10 ** 6; // 2 USDC
+        uint256 daiAmount = 2 * 10 ** 18; // 2 DAI
         uint256 minUsdcAmount = (usdcAmount * 95) / 100;
         uint256 minDaiAmount = (daiAmount * 95) / 100;
 
@@ -53,5 +53,4 @@ contract UniswapTest {
         //     target
         // );
     }
-
 }
